+++ src/pages/staking/staking.page.vue
@@ -28,7 +28,7 @@
                             <span class="d-block">Delegated Stake Fee {{delegateFeePercentage}} %</span>
                         </template>
 
-                        <div class="pt-4" v-if="account.plainAccount">
+                        <div class="pt-4" v-if="account && account.plainAccount">
                             <span class="fw-bold fs-0">Claimable</span>
                             <balance :key="`delegated-balance`"  :balance="account.plainAccount.claimable" asset=""></balance>
                         </div>
